add zellij config file. moved nvim config with lazy to the folder nvim-lazy. Now the folder nvim has the custom config again
