--- alacritty/alacritty.yml
@@ -1,64 +1,0 @@
-shell:
-  program: /usr/bin/zsh
-
-window:
-  padding:
-    x: 0
-    y: 0
-  class:
-    instance: Alacritty
-    general: Alacritty
-
-scrolling:
-  history: 10000
-  multiplier: 3
-
-font:
-  normal:
-    family: JetBrains Mono Nerd Font
-    style: Medium
-  bold:
-    family: JetBrains Mono Nerd Font
-    style: Bold
-  italic:
-    family: JetBrains Mono Nerd Font
-    style: MediumItalic
-  bold_italic:
-    family: JetBrains Mono Nerd Font
-    style: BoldItalic
-  size: 12
-draw_bold_text_with_bright_colors: true
-
-# TokyoNight Alacritty Colors
-colors:
-  # Default colors
-  primary:
-    background: '0x1a1b26'
-    foreground: '0xc0caf5'
-
-  # Normal colors
-  normal:
-    black:   '0x15161e'
-    red:     '0xf7768e'
-    green:   '0x9ece6a'
-    yellow:  '0xe0af68'
-    blue:    '0x7aa2f7'
-    magenta: '0xbb9af7'
-    cyan:    '0x7dcfff'
-    white:   '0xa9b1d6'
-
-  # Bright colors
-  bright:
-    black:   '0x414868'
-    red:     '0xf7768e'
-    green:   '0x9ece6a'
-    yellow:  '0xe0af68'
-    blue:    '0x7aa2f7'
-    magenta: '0xbb9af7'
-    cyan:    '0x7dcfff'
-    white:   '0xc0caf5'
-
-  indexed_colors:
-    - { index: 16, color: '0xff9e64' }
-    - { index: 17, color: '0xdb4b4b' }
-
