--- nvim-custom/init.lua
@@ -1,3 +1,0 @@
-require("bocha13.settings")
-require("bocha13.lazy")
-require("bocha13.keybindings")
