--- nvim-custom/lua/bocha13/icons.lua
@@ -1,37 +1,0 @@
-Icons = {
-  Array = " ",
-  Boolean = " ",
-  Class = " ",
-  Color = " ",
-  Constant = " ",
-  Constructor = " ",
-  Copilot = " ",
-  Enum = " ",
-  EnumMember = " ",
-  Event = " ",
-  Field = " ",
-  File = " ",
-  Folder = " ",
-  Function = " ",
-  Interface = " ",
-  Key = " ",
-  Keyword = " ",
-  Method = " ",
-  Module = " ",
-  Namespace = " ",
-  Null = " ",
-  Number = " ",
-  Object = " ",
-  Operator = " ",
-  Package = " ",
-  Property = " ",
-  Reference = " ",
-  Snippet = " ",
-  String = " ",
-  Struct = " ",
-  Text = " ",
-  TypeParameter = " ",
-  Unit = " ",
-  Value = " ",
-  Variable = " ",
-}
