--- nvim-custom/lua/bocha13/plugins/init.lua
@@ -1,4 +1,0 @@
-return {
-	"christoomey/vim-tmux-navigator",
-	"ThePrimeagen/vim-be-good",
-}
