--- nvim-custom/lua/bocha13/plugins/mini.lua
@@ -1,31 +1,0 @@
-return {
-	{ "echasnovski/mini.nvim", version = false },
-	-- auto pairs
-	{
-		"echasnovski/mini.pairs",
-		event = "VeryLazy",
-		config = function(_, opts)
-			require("mini.pairs").setup(opts)
-		end,
-	},
-	-- buffer remove
-	{
-		"echasnovski/mini.bufremove",
-		keys = {
-			{
-				"<A-c>",
-				function()
-					require("mini.bufremove").delete(0, false)
-				end,
-				desc = "Delete Buffer",
-			},
-			{
-				"<leader>bD",
-				function()
-					require("mini.bufremove").delete(0, true)
-				end,
-				desc = "Delete Buffer (Force)",
-			},
-		},
-	},
-}
