--- nvim-custom/lua/bocha13/plugins/undotree.lua
@@ -1,7 +1,0 @@
-return {
-  "mbbill/undotree",
-  cmd = "UndotreeToggle",
-  keys = {
-    { "<leader>u", vim.cmd.UndotreeToggle, desc = "Toggle undotree" }
-  }
-}
