+++ nvim-lazy/init.lua
@@ -1,0 +1,2 @@
+-- bootstrap lazy.nvim, LazyVim and your plugins
+require("config.lazy")
