+++ nvim-lazy/lua/config/autocmds.lua
