+++ nvim-lazy/lua/config/options.lua
