+++ nvim-lazy/lua/plugins/cmp.lua
@@ -50,6 +50,7 @@ return {
           }),
         }),
         sources = cmp.config.sources({
+          { name = "copilot" },
           { name = "nvim_lsp" },
           { name = "luasnip" },
           { name = "buffer" },
