+++ nvim-lazy/lua/plugins/core.lua
