+++ nvim-lazy/lua/plugins/disable.lua
