+++ nvim-lazy/lua/plugins/ia.lua
@@ -14,16 +14,4 @@ return {
     "zbirenbaum/copilot-cmp",
     opts = {},
   },
-  {
-    "jackMort/ChatGPT.nvim",
-    event = "VeryLazy",
-    config = function()
-      require("chatgpt").setup()
-    end,
-    dependencies = {
-      "MunifTanjim/nui.nvim",
-      "nvim-lua/plenary.nvim",
-      "nvim-telescope/telescope.nvim",
-    },
-  },
 }
