+++ nvim/init.lua
@@ -1,2 +1,3 @@
--- bootstrap lazy.nvim, LazyVim and your plugins
+require("config.keymaps")
+require("config.settings")
 require("config.lazy")
