+++ nvim/lua/config/settings.lua
@@ -4,15 +4,15 @@ vim.opt.guicursor = ""
 vim.g.loaded_netrw = 1
 vim.g.loaded_netrwPlugin = 1
 
--- vim.api.nvim_command("set fillchars+=vert:\\▎")
+vim.api.nvim_command("set fillchars+=vert:\\▎")
 
 -- remove auto comment on new line after comment
 vim.api.nvim_create_autocmd("BufEnter", {
-	callback = function()
-		vim.opt.formatoptions:remove({ "c", "r", "o" })
-	end,
-	group = general,
-	desc = "Disable New Line Comment",
+  callback = function()
+    vim.opt.formatoptions:remove({ "c", "r", "o" })
+  end,
+  group = general,
+  desc = "Disable New Line Comment",
 })
 
 vim.opt.showmode = false
@@ -49,7 +49,7 @@ vim.g.mapleader = " "
 
 -- Format on save
 vim.api.nvim_create_autocmd("BufWritePre", {
-	callback = function()
-		vim.lsp.buf.format()
-	end,
+  callback = function()
+    vim.lsp.buf.format()
+  end,
 })
