+++ nvim/lua/plugins/editor.lua
@@ -1,17 +1,9 @@
 return {
   {
     "numToStr/Comment.nvim",
+    event = { "BufReadPre", "BufNewFile" },
     keys = { { "gcc" }, { "gbc" }, { "gc", mode = "v" }, { "gb", mode = "v" } },
     config = true
-  },
-  {
-    "tpope/vim-fugitive",
-    cmd = "Git",
-    keys = {
-      { "<leader>gc", ":Git commit -m \"",            desc = "Git commit with message" },
-      { "<leader>gp", ":Git push -u origin HEAD<CR>", desc = "Git push to origin" },
-      { "<leader>gs", ":Git<CR>",                     desc = "Open fugitive panel" },
-    }
   },
   {
     "kylechui/nvim-surround",
@@ -26,8 +18,4 @@ return {
     dependencies = { "nvim-web-devicons" },
     event = "VeryLazy"
   },
-  {
-    "windwp/nvim-ts-autotag",
-    event = "InsertEnter"
-  }
 }
