+++ nvim/lua/plugins/init.lua
@@ -1,0 +1,3 @@
+return {
+  "christoomey/vim-tmux-navigator",
+}
