+++ nvim/lua/plugins/lsp/init.lua
@@ -1,0 +1,1 @@
+return {}
