+++ nvim/lua/plugins/lsp/lspconfig.lua
@@ -1,0 +1,94 @@
+return {
+  "neovim/nvim-lspconfig",
+  event = { "BufReadPre", "BufNewFile" },
+  dependencies = {
+    "hrsh7th/cmp-nvim-lsp",
+    { "antosha417/nvim-lsp-file-operations", config = true }
+  },
+  config = function()
+    local lspconfig = require("lspconfig")
+    local cmp_nvim_lsp = require("cmp_nvim_lsp")
+
+    -- set keymaps
+    local keymap = vim.keymap
+    local on_attach = function(client, bufnr)
+      local opts = { noremap = true, silent = true, buffer = bufnr }
+
+      -- disable tsserver formatting when eslint is enabled
+      if client.name == "eslint" then
+        client.server_capabilities.documentFormattingProvider = true
+      elseif client.name == "tsserver" then
+        client.server_capabilities.documentFormattingProvider = false
+      end
+
+      keymap.set("n", "gR", "<cmd>Telescope lsp_references<CR>", opts)
+      keymap.set("n", "gD", vim.lsp.buf.declaration, opts)
+      keymap.set("n", "gd", "<cmd>Telescope lsp_definitions<CR>", opts)
+      keymap.set("n", "gi", "<cmd>Telescope lsp_implementations<CR>", opts)
+      keymap.set("n", "gt", "<cmd>Telescope lsp_type_definitions<CR>", opts)
+      keymap.set('n', 'gr', vim.lsp.buf.references, opts)
+      keymap.set("n", "K", vim.lsp.buf.hover, opts)
+      keymap.set("i", "<C-k>", vim.lsp.buf.signature_help, opts)
+      keymap.set({ "n", "v" }, "<leader>ca", vim.lsp.buf.code_action, opts)
+      keymap.set("n", "<leader>rn", vim.lsp.buf.rename, opts)
+      keymap.set("n", "<space>cd", vim.diagnostic.open_float, opts)
+    end
+
+    -- used to enable autocompletion
+    local capabilities = cmp_nvim_lsp.default_capabilities()
+    local signs = { Error = "E", Warn = "W", Hint = "H", Info = "I" }
+    for type, icon in pairs(signs) do
+      local hl = "DiagnosticSign" .. type
+      vim.fn.sign_define(hl, { text = icon, texthl = hl, numhl = "" })
+    end
+
+    --configure servers
+    lspconfig["html"].setup({
+      capabilities = capabilities,
+      on_attach = on_attach,
+    })
+
+    lspconfig["lua_ls"].setup({
+      capabilities = capabilities,
+      on_attach = on_attach,
+      settings = {
+        Lua = {
+          diagnostics = {
+            globals = { "vim", "general" }
+          },
+          workspace = {
+            library = {
+              [vim.fn.expand("$VIMRUNTIME/lua")] = true,
+              [vim.fn.stdpath("config") .. "/lua"] = true,
+            }
+          }
+        }
+      }
+    })
+
+    lspconfig["tsserver"].setup({
+      capabilities = capabilities,
+      on_attach = on_attach,
+    })
+
+    lspconfig["tailwindcss"].setup({
+      capabilities = capabilities,
+      on_attach = on_attach,
+    })
+
+    lspconfig["cssls"].setup({
+      capabilities = capabilities,
+      on_attach = on_attach,
+    })
+
+    lspconfig["graphql"].setup({
+      capabilities = capabilities,
+      on_attach = on_attach,
+    })
+
+    lspconfig["eslint"].setup({
+      capabilities = capabilities,
+      on_attach = on_attach,
+    })
+  end
+}
