--- warp/themes/catppuccin-mocha.yaml
@@ -1,23 +1,0 @@
-foreground: "#cdd6f4"
-accent: "#f2cdcd"
-background: "#191724"
-terminal_colors:
-  bright:
-    blue: "#74c7ec"
-    red: "#eba0ac"
-    green: "#94e2d5"
-    black: "#313244"
-    yellow: "#f9e2af"
-    cyan: "#74c7ec"
-    magenta: "#f5c2e7"
-    white: "#cdd6f4"
-  normal:
-    magenta: "#cba6f7"
-    yellow: "#f9e2af"
-    white: "#cdd6f4"
-    blue: "#89b4fa"
-    black: "#191724"
-    red: "#f38ba8"
-    green: "#a6e3a1"
-    cyan: "#89dceb"
-details: "darker"
