--- warp/themes/warp-tokyonight.yaml
@@ -1,23 +1,0 @@
-accent: "#65bcff"
-background: "#222436"
-foreground: "#ffffff"
-details: "darker"
-terminal_colors:
-  normal:
-    black: "#1e2030"
-    red: "#ff757f"
-    green: "#c3e88d"
-    yellow: "#ffc777"
-    blue: "#82aaff"
-    magenta: "#c099ff"
-    cyan: "#86e1fc"
-    white: "#f1f1f1"
-  bright:
-    black: "#8e8e8e"
-    red: "#ffc4bd"
-    green: "#d6fcb9"
-    yellow: "#fefdd5"
-    blue: "#c1e3fe"
-    magenta: "#ffb1fe"
-    cyan: "#e5e6fe"
-    white: "#feffff"
